refactor: fix consistency between commonStyles
