+++ src/commonComponents/Footer/Cover.js
@@ -3,10 +3,10 @@ import styled from "styled-components";
 export default styled.div`
   width: 64px;
   height: 88px;
-  margin: 14px 10px 14px;
-  border-radius: 2px;
-  box-shadow: 0px 2px 4px 0px #0000001A;
+  border-radius: 4px;
+  margin: 14px 10px;
   background-color: #FFF;
+  box-shadow: 0px 2px 4px 0px #0000001A;
   display: flex;
   align-items: center;
   justify-content: center;
