+++ src/commonComponents/Header/Header.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useHistory } from "react-router";
 import { useLocation } from "react-router-dom";
 
-import Wrapper from "./styles/Wrapper";
+import Wrapper from "./Wrapper";
 
 export default function Header() {
   const history = useHistory();
