+++ src/commonStyles/MovieCover.js
@@ -1,7 +1,8 @@
 import styled from "styled-components";
 
 export default styled.img`
-  width: ${(props) => (props.footer ? '48px' : '128px')};
-  height: ${(props) => (props.footer ? '72px' : '192px')};
+  width: ${(props) => (props.footer ? '56px' : '136px')};
+  height: ${(props) => (props.footer ? '84px' : '200px')};
+  border-radius: 4px;
   object-fit: cover;
 `;
